Move consistency checking to cacher subpackage
