+++ staging/src/k8s.io/apiserver/pkg/storage/cacher/delegator.go
@@ -18,25 +18,17 @@ package cacher
 
 import (
 	"context"
-	"fmt"
-	"hash"
-	"hash/fnv"
-	"os"
-	"strconv"
 	"sync"
 	"time"
 
 	"go.opentelemetry.io/otel/attribute"
 	"k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/api/meta"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/runtime/schema"
-	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/apimachinery/pkg/watch"
 	"k8s.io/apiserver/pkg/audit"
 	"k8s.io/apiserver/pkg/features"
 	"k8s.io/apiserver/pkg/storage"
+	"k8s.io/apiserver/pkg/storage/cacher/consistency"
 	"k8s.io/apiserver/pkg/storage/cacher/delegator"
 	"k8s.io/apiserver/pkg/storage/cacher/metrics"
 	"k8s.io/apiserver/pkg/storage/cacher/store"
@@ -45,32 +37,18 @@ import (
 	"k8s.io/klog/v2"
 )
 
-var (
-	// ConsistencyCheckPeriod is the period of checking consistency between etcd and cache.
-	// 5 minutes were proposed to match the default compaction period. It's magnitute higher than
-	// List latency SLO (30 seconds) and timeout (1 minute).
-	ConsistencyCheckPeriod = 5 * time.Minute
-	// panicOnCacheInconsistency enables the consistency checking mechanism for cache.
-	// Based on KUBE_WATCHCACHE_CONSISTENCY_CHECKER environment variable.
-	panicOnCacheInconsistency = false
-)
-
-func init() {
-	panicOnCacheInconsistency, _ = strconv.ParseBool(os.Getenv("KUBE_WATCHCACHE_CONSISTENCY_CHECKER"))
-}
-
 func NewCacheDelegator(cacher *Cacher, storage storage.Interface) *CacheDelegator {
 	d := &CacheDelegator{
 		cacher:  cacher,
 		storage: storage,
 		stopCh:  make(chan struct{}),
 	}
-	if utilfeature.DefaultFeatureGate.Enabled(features.DetectCacheInconsistency) || panicOnCacheInconsistency {
-		d.checker = newConsistencyChecker(cacher.resourcePrefix, cacher.groupResource, cacher.newListFunc, cacher, storage)
+	if utilfeature.DefaultFeatureGate.Enabled(features.DetectCacheInconsistency) || consistency.PanicOnCacheInconsistency {
+		d.checker = consistency.NewChecker(cacher.resourcePrefix, cacher.groupResource, cacher.newListFunc, cacher, storage)
 		d.wg.Add(1)
 		go func() {
 			defer d.wg.Done()
-			d.checker.startChecking(d.stopCh)
+			d.checker.Run(d.stopCh)
 		}()
 	}
 	return d
@@ -79,7 +57,7 @@ func NewCacheDelegator(cacher *Cacher, storage storage.Interface) *CacheDelegato
 type CacheDelegator struct {
 	cacher  *Cacher
 	storage storage.Interface
-	checker *consistencyChecker
+	checker *consistency.Checker
 
 	wg       sync.WaitGroup
 	stopOnce sync.Once
@@ -285,216 +263,4 @@ func (c *CacheDelegator) Stop() {
 		close(c.stopCh)
 	})
 	c.wg.Wait()
-}
-
-func newConsistencyChecker(resourcePrefix string, groupResource schema.GroupResource, newListFunc func() runtime.Object, cacher cacher, etcd getLister) *consistencyChecker {
-	return &consistencyChecker{
-		groupResource:  groupResource,
-		resourcePrefix: resourcePrefix,
-		newListFunc:    newListFunc,
-		cacher:         cacher,
-		etcd:           etcd,
-	}
-}
-
-type consistencyChecker struct {
-	resourcePrefix string
-	groupResource  schema.GroupResource
-	newListFunc    func() runtime.Object
-
-	cacher cacher
-	etcd   getLister
-}
-
-type cacher interface {
-	getLister
-	Ready() bool
-	MarkConsistent(bool)
-}
-
-type getLister interface {
-	GetList(ctx context.Context, key string, opts storage.ListOptions, listObj runtime.Object) error
-}
-
-func (c consistencyChecker) startChecking(stopCh <-chan struct{}) {
-	klog.V(3).InfoS("Cache consistency check start", "group", c.groupResource.Group, "resource", c.groupResource.Resource)
-	jitter := 0.5 // Period between [interval, interval * (1.0 + jitter)]
-	sliding := true
-	// wait.JitterUntilWithContext starts work immediately, so wait first.
-	select {
-	case <-time.After(wait.Jitter(ConsistencyCheckPeriod, jitter)):
-	case <-stopCh:
-	}
-	wait.JitterUntilWithContext(wait.ContextForChannel(stopCh), c.check, ConsistencyCheckPeriod, jitter, sliding)
-}
-
-func (c *consistencyChecker) check(ctx context.Context) {
-	digests, err := c.calculateDigests(ctx)
-	if err != nil {
-		klog.ErrorS(err, "Cache consistency check error", "group", c.groupResource.Group, "resource", c.groupResource.Resource)
-		metrics.StorageConsistencyCheckTotal.WithLabelValues(c.groupResource.Group, c.groupResource.Resource, "error").Inc()
-		return
-	}
-	if digests.CacheDigest == digests.EtcdDigest {
-		klog.V(3).InfoS("Cache consistency check passed", "group", c.groupResource.Group, "resource", c.groupResource.Resource, "resourceVersion", digests.ResourceVersion, "digest", digests.CacheDigest)
-		metrics.StorageConsistencyCheckTotal.WithLabelValues(c.groupResource.Group, c.groupResource.Resource, "success").Inc()
-		c.cacher.MarkConsistent(true)
-		return
-	}
-	klog.ErrorS(nil, "Cache consistency check failed", "group", c.groupResource.Group, "resource", c.groupResource.Resource, "resourceVersion", digests.ResourceVersion, "etcdDigest", digests.EtcdDigest, "cacheDigest", digests.CacheDigest, "diffDetail", digests.DiffDetail)
-	metrics.StorageConsistencyCheckTotal.WithLabelValues(c.groupResource.Group, c.groupResource.Resource, "failure").Inc()
-	if panicOnCacheInconsistency {
-		panic(fmt.Sprintf("Cache consistency check failed, group: %q, resource: %q, resourceVersion: %q, etcdDigest: %q, cacheDigest: %q, diffDetail: %v", c.groupResource.Group, c.groupResource.Resource, digests.ResourceVersion, digests.EtcdDigest, digests.CacheDigest, digests.DiffDetail))
-	}
-	c.cacher.MarkConsistent(false)
-}
-
-func (c *consistencyChecker) calculateDigests(ctx context.Context) (*storageDigest, error) {
-	if !c.cacher.Ready() {
-		return nil, fmt.Errorf("cache is not ready")
-	}
-	// variables for tracking the diff in consistency checker
-	cacheItems := []namespaceNameRV{}
-	var foundDiff *diffDetail
-	var etcdIndex int
-
-	cacheDigest, cacheResourceVersion, err := c.calculateStoreDigest(ctx, c.cacher, "0", 0, func(obj metav1.Object) {
-		// collect items from cacher's list
-		cacheItems = append(cacheItems, namespaceNameRV{Namespace: obj.GetNamespace(), Name: obj.GetName(), RV: obj.GetResourceVersion()})
-	})
-	if err != nil {
-		return nil, fmt.Errorf("failed calculating cache digest: %w", err)
-	}
-	etcdDigest, etcdResourceVersion, err := c.calculateStoreDigest(ctx, c.etcd, cacheResourceVersion, storageWatchListPageSize, func(obj metav1.Object) {
-		// compare the item in etcd's list against cacheItems which is cacher's list
-		if foundDiff != nil {
-			return
-		}
-		etcdItem := namespaceNameRV{Namespace: obj.GetNamespace(), Name: obj.GetName(), RV: obj.GetResourceVersion()}
-		if len(cacheItems) <= etcdIndex {
-			foundDiff = &diffDetail{Index: etcdIndex, EtcdItem: &etcdItem}
-			cacheItems = nil // don't need it any more and nil it to allow GC to collect it
-			return
-		}
-		if cacheItem := cacheItems[etcdIndex]; cacheItem != etcdItem {
-			foundDiff = &diffDetail{Index: etcdIndex, EtcdItem: &etcdItem, CacheItem: &cacheItem}
-			cacheItems = nil // don't need it any more and nil it to allow GC to collect it
-			return
-		}
-		etcdIndex += 1
-	})
-	if err != nil {
-		return nil, fmt.Errorf("failed calculating etcd digest: %w", err)
-	}
-	if len(cacheItems) > etcdIndex {
-		cacheItem := cacheItems[etcdIndex]
-		foundDiff = &diffDetail{Index: etcdIndex, CacheItem: &cacheItem}
-	}
-	if cacheResourceVersion != etcdResourceVersion {
-		return nil, fmt.Errorf("etcd returned different resource version then expected, cache: %q, etcd: %q", cacheResourceVersion, etcdResourceVersion)
-	}
-	return &storageDigest{
-		ResourceVersion: cacheResourceVersion,
-		CacheDigest:     cacheDigest,
-		EtcdDigest:      etcdDigest,
-		DiffDetail:      foundDiff,
-	}, nil
-}
-
-type namespaceNameRV struct {
-	Namespace string
-	Name      string
-	RV        string
-}
-
-type diffDetail struct {
-	Index     int
-	CacheItem *namespaceNameRV
-	EtcdItem  *namespaceNameRV
-}
-
-type storageDigest struct {
-	ResourceVersion string
-	CacheDigest     string
-	EtcdDigest      string
-	DiffDetail      *diffDetail
-}
-
-func (c *consistencyChecker) calculateStoreDigest(ctx context.Context, store getLister, resourceVersion string, limit int64, metaVisitor func(objectMeta metav1.Object)) (digest, rv string, err error) {
-	opts := storage.ListOptions{
-		Recursive:       true,
-		Predicate:       storage.Everything,
-		ResourceVersion: resourceVersion,
-	}
-	opts.Predicate.Limit = limit
-	if resourceVersion == "0" {
-		opts.ResourceVersionMatch = metav1.ResourceVersionMatchNotOlderThan
-	} else {
-		opts.ResourceVersionMatch = metav1.ResourceVersionMatchExact
-	}
-	h := fnv.New64()
-	for {
-		resp := c.newListFunc()
-		err = store.GetList(ctx, c.resourcePrefix, opts, resp)
-		if err != nil {
-			return "", "", err
-		}
-		err = addListToDigest(h, resp, metaVisitor)
-		if err != nil {
-			return "", "", err
-		}
-		list, err := meta.ListAccessor(resp)
-		if err != nil {
-			return "", "", err
-		}
-		if resourceVersion == "0" {
-			resourceVersion = list.GetResourceVersion()
-		}
-		continueToken := list.GetContinue()
-		if continueToken == "" {
-			return fmt.Sprintf("%x", h.Sum64()), resourceVersion, nil
-		}
-		opts.Predicate.Continue = continueToken
-		opts.ResourceVersion = ""
-		opts.ResourceVersionMatch = ""
-	}
-}
-
-func addListToDigest(h hash.Hash64, list runtime.Object, metaVisitor func(objectMeta metav1.Object)) error {
-	return meta.EachListItem(list, func(obj runtime.Object) error {
-		objectMeta, err := meta.Accessor(obj)
-		if err != nil {
-			return err
-		}
-		err = addObjectToDigest(h, objectMeta)
-		if err != nil {
-			return err
-		}
-		metaVisitor(objectMeta)
-		return nil
-	})
-}
-
-func addObjectToDigest(h hash.Hash64, objectMeta metav1.Object) error {
-	_, err := h.Write([]byte(objectMeta.GetNamespace()))
-	if err != nil {
-		return err
-	}
-	_, err = h.Write([]byte("/"))
-	if err != nil {
-		return err
-	}
-	_, err = h.Write([]byte(objectMeta.GetName()))
-	if err != nil {
-		return err
-	}
-	_, err = h.Write([]byte("/"))
-	if err != nil {
-		return err
-	}
-	_, err = h.Write([]byte(objectMeta.GetResourceVersion()))
-	if err != nil {
-		return err
-	}
-	return nil
 }
